LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/custom-element.window-expected.txt


PASS document.open() works in custom element constructor for createElement()
PASS document.open() is forbidden in custom element constructor when creating element from parser
